[CORDA-1971]: Switched exception error codes to use Base36 encoding.
